==> 1_data_preprocessing.ipynb <==
# Yelp Dataset Exploration

   This Jupyter notebook is designed to explore the Yelp Dataset. It involves loading several JSON datasets into pandas DataFrames, preprocessing, merging, and exporting them for further analysis. Our goal is to understand the characteristics of businesses, users, reviews, and photos within a specific geographical location and domain (restaurants).


import pandas as pd
import numpy as np


## Loading Datasets

The first step in our exploration is to load various datasets provided by Yelp. These datasets include information on businesses, users, reviews, and photos. We use the `pd.read_json` function to load each dataset in chunks for efficient memory management.




# Define the path to your JSON file
file_path = "./dataset/jsons/yelp_academic_dataset_business.json"

# Directly concatenate the chunks read from the JSON file into a single DataFrame
business = pd.concat(
    pd.read_json(file_path, lines=True, chunksize=10000),
    ignore_index=True
)

# 'business' DataFrame now contains all the data from the JSON file




# Define the path to your JSON file
file_path = "./dataset/jsons/yelp_academic_dataset_user.json"

# Directly concatenate the chunks read from the JSON file into a single DataFrame
user = pd.concat(
    pd.read_json(file_path, lines=True, chunksize=10000),
    ignore_index=True
)

# The 'user' DataFrame now contains all the data from the JSON file




# Define the path to your JSON file
file_path = "./dataset/jsons/yelp_academic_dataset_review.json"

# Read the JSON file in chunks and concatenate directly without intermediate list comprehension
review = pd.concat(
    pd.read_json(file_path, lines=True,
                 dtype={'review_id': str, 'user_id': str,
                        'business_id': str, 'stars': 'int8',
                        'date': str, 'text': str, 'useful': 'int8',
                        'funny': 'int8', 'cool': 'int8'},
                 chunksize=10000),
    ignore_index=True
)

# Now 'review' contains the entire dataset




# Define the path to your JSON file
file_path = "./dataset/jsons/photos.json"

# Directly concatenate the chunks read from the JSON file into a single DataFrame
photo = pd.concat(
    pd.read_json(file_path, lines=True, chunksize=10000),
    ignore_index=True
)

# The 'photo' DataFrame now contains all the data from the JSON file

## Data Preprocessing

After loading the data, we focus on preprocessing. This includes filtering datasets based on specific criteria (like selecting only Nevada-based businesses), merging datasets, and calculating new metrics (such as an adjusted score for businesses based on their reviews).


# Group by 'business_id' and take the first 'photo_id' from each group
first_photo = photo.groupby('business_id', as_index=False).first()[['business_id', 'photo_id']]


list_of_states = ['NV']
business = business[business.state.isin(list_of_states)]

import pandas as pd
import numpy as np

def filter_restaurant_businesses(df):
    restaurant_set = set(['restaurants', 'fast food', 'sandwiches', 'caterers', 'deserts', 'burgers'])
    
    # Function to determine if a business is a restaurant
    def is_restaurant(categories):
        if pd.isna(categories):
            return False
        categories_set = set(map(str.strip, map(str.lower, categories.split(','))))
        if categories_set.intersection(restaurant_set):
            return True
        return False  # Assuming businesses not explicitly identified as restaurants are not restaurants
    
    mask = df['categories'].apply(is_restaurant)
    
    return df[mask]

business = filter_restaurant_businesses(business)


import pandas as pd
import numpy as np

# Assuming 'business' is your DataFrame and already defined.
# Perform the merge operation and assign the result back to the original DataFrame variable
business = pd.merge(business, first_photo, on="business_id", how="left")

# Fill NaN values in the 'photo_id' column with 'no-image'
business['photo_id'] = business['photo_id'].fillna('no-image')


# Calculate the global mean rating across all businesses
global_mean = (business['stars'] * business['review_count']).sum() / business['review_count'].sum()
# Determine the median review count (50th percentile) dynamically
k = business['review_count'].quantile(0.5)
# Calculate the adjusted score for each business
business['adjusted_score'] = ((business['review_count'] * business['stars']) + (k * global_mean)) / (business['review_count'] + k)
# Sorting the 'business' DataFrame by 'adjusted_score' in descending order, in place
business.sort_values(by='adjusted_score', ascending=False, inplace=True)
# reset index
business.reset_index(drop=True, inplace=True)
print(business.info())




user = user[(user['review_count'] > 0) & (user['average_stars'] != 0)].reset_index(drop=True)
print(user.info())



unique_user_ids = set(user.user_id.unique())
unique_business_ids = set(business.business_id.unique())

review = review[(review.user_id.isin(unique_user_ids)) & 
                (review.business_id.isin(unique_business_ids)) & 
                (review.stars != 0)]

review.reset_index(drop=True, inplace=True)
print(review.info())


photo = photo[photo.business_id.isin(unique_business_ids)]
photo.reset_index(drop=True, inplace=True)
print(photo.info())

## Exporting Data

With the data now preprocessed, the final step involves exporting the modified DataFrames to both CSV and pickle formats for easy access in future analyses.


business.to_csv(path_or_buf='business.csv',index=False)
user.to_csv(path_or_buf='user.csv',index=False)
review.to_csv(path_or_buf='review.csv',index=False)
photo.to_csv(path_or_buf='photo.csv',index=False)

business.to_pickle('business.pkl')
user.to_pickle('user.pkl')
review.to_pickle('review.pkl')
photo.to_pickle('photo.pkl')

review[['user_id', 'business_id', 'stars', 'date']].to_pickle('review_cache.pkl')
